requirements.txt
pymupdf
google-genai
python-dotenv
fastapi
uvicorn
python-multipart
uvicorn
main:app --reload --port 8000
